[app] Improve instrumentation middleware
We unified the exported function name across the httpmetrics, httplog
and httptracer middleware. They all export a Handler function which can
be used in the router.

The logging and tracing middleware also catchs panics now to write a log
line / create a span.

Besides that we also fixed a small bug caused by the notification
drawer, introduced in #389. Because of the notification drawer the
drawer in the PageSectionContent component wasn't using it's full height
anymore, which caused problems for the select boxes in the toolbar and
the pages showing graphs via cytoscape. This should be fixed now.