[rss] Display RSS Feeds in Notifications (#392)
It is now possible to use the RSS plugin to display the items from an
RSS feed within the notifications. RSS feed items are always marked as
unread and info.

We also added a page implementation for the plugin to show the items of
a single RSS feed. This allows us to link from a notification entry to
the complete rss feed.