[app] Display Kubernetes Resources in Notifications
It is now possible to display Kubernetes Resources within the
notifications. For this the existing "resources" plugin with type "app"
can be used. It allows the same options as when the plugin is used
within a panel.

We also added a new "filter" option for the resources plugin, which
allows users to filter the retrieved list of resources via JSONPath.
With this new option it is possible to only display unhealthy Pods in
the notifications.