[app] Custom User and Team Notifications
It is now possible to customize the notification settings per user and
team. For that a new "notifications" property was added to both CRDs,
which can be used to define custom notification groups. If a user or
teams has not defined any notifications the global notification groups
are used.

We also adjusted the documentation and added a "Notification Options"
section for each plugin.

Besides that we also fixed a bug in the applications and teams API,
where the "Teams" field was used to check the users permissions instead
of the "Permissions.Teams" field from the user context. This was
noticed, because we added the list of teams from the corresponding
header also when authentication is disabled.