[app] Fix Permission Handling for Apps and Teams
This commit fixes a bug in the permission handling for applications and
teams. Because we not only checked the "all" query parameter in the
corresponding routes, it could happen that we return all applications /
teams, when a user just wants to see his own applications / teams.

We are now just checking if the "all" query parameter is true to set the
teams to "nil". For all other use cases we are directly using the teams
form the authenticated user object.

When authentication is disabled or when no team header is present the
teams for the user are automatically set to "nil" so that the other
checks are not necessary anymore.