[app] Fix Shutdown Logic (#396)
We forgot to shutdown the created tracer provider. This is now
implemented, by returning a client in the tracer setup. This client
exposes a Shutdown method which can be used to gracefully shutdown the
tracer provider.

We also fixed the context timeouts for all http server, so that the hub,
satellite and app server are having a timeout of 30 seconds and the
metrics server has a timeout of 10 seconds as it is also used for the
tracer client.

This way we always stay within the 90 seconds timeout, which was
configured at the Kubernetes deployments for the hub and satellite.