[github] Add GitHub Plugin (#398)
Add a GitHub plugin, which can be used to view issues and pull requests
for a repository and organization. It is also possible to list
repositories and members of an organization and team. Users of kobs can
use the "userpullrequests" panel to add there created, assigned,
mentioned or review requested pull requests to their profile page.

The plugin requires a GitHub OAuth App to authorize users to access the
organizations data via the GitHub API.