[github] Add Notifications
It is now possible to use the GitHub plugin for notifications. This
means it is possible to show a users notifications or a users created,
mentioned, assigned or review requested pull requests in the
notification drawer.

We also added a new panel "usernotifications" to show the users GitHub
notifications within a dashboard (e.g. on his profile page).

Last but not least we renamed the cookie used for authentication against
the GitHub API. This was necessary, because until now the cookie always
had the same name which doesn't allow multiple instances of the GitHub
plugin. Now we add the configured organization name to the cookie name,
so that multiple instances of the plugin can be used.