[app] Add Integrations for Kubernetes Resources (#401)
Add a new integrations feature for kobs. In this first iteration we
implemented the suggested integrations feature from #255 for Kubernetes
Resources.

This integration allows an administrator of kobs to add a set of default
dashboards to Kubernetes Resources. This way we can add for example a
resource usage dashboard to all Pods, without that each Pod must have
the "kobs.io/dashboards" annotation.

The configured dashboards in the integrations and the dashboards from
the "kobs.io/dashboards" annotation are always merged and not
overwritte.