[app] Add "mongodb" driver (#402)
Add a new "mongodb" driver to use MongoDB as storage backend instead of
Bolt. The MongoDB storage backend can be used to scale kobs, when the
Bolt backend reaches it's limits.

The MongoDB storage backend can be used via the "hub.store.driver" flag.
To get the best results with the MongoDB storage backend we recommend to
use two deployments for the hub: One deployment can be used to serve the
API and React UI ("--hub.mode=server") and the other one can be used to
sync the data from the configured satellites and write the data into
MongoDB ("--hub.mode=watcher"). This deployment architecture is also
supported by the Helm chart, via the "watcher.enabled" value.

To add a readiness and liveness probe for the watcher Pods, the metrics
server has now a "/health" endpoint which is used in the Helm chart. We
also had to update the Go dependencies, to use the new MongoDB storage
backend.