[app] Add missing sort keys to MongoDB queries
We missed the sort key for multiple queries during the implementation of
the MongoDB driver. These keys are now added so that the order is the
exact same as it is for the Bolt driver.

We also added a "make vet" command to the continuous integration
pipeline to check for suspicious constructs like the following which was
introduced by the MongoDB driver:

go.mongodb.org/mongo-driver/bson/primitive.E composite literal uses unkeyed fields

This is now fixed by changing all occurences of "bson.D{{"key", "value"}}"
to "bson.D{{Key: "key", Value: "value"}}".