[sql] Add "singlestats" Chart (#407)
The SQL plugin contains a new "singlestats" chart, which can be selected
in the panel options. This chart type can be used to render one or
multiple single values returned by a query. It is also possible to
specify a list of thresholds for the singlestats chart, to set the
background color of the chart based on the value / thresholds.