[app] Improve Usability
Improve usability for end users by showing a hint when no applications
are returned for a users search and when he selected the "own" filter.
Within the info message a user can now also easily switch to the all
filter.

The order of the tabs for resources was always random (depennding on
which API call was finished first). Now the tabs are sorted
alphabatically, so that a user can always expect the same order.

We also fixed a bug, were the user did not receive a proper error
message when a selected resource id was invalid. Now we do not return a
incorrect JSON response, so that we can render a proper error message.