[app] Fix 404 error for ACE worker files (#410)
When a user opens a page with an Editor component a 404 error was thrown
in the console, because the worker files for ACE (the underlying editor)
could not be found. This is fixed by disabling workers using the
"useWorker: false" in the "setOptions" property.

This also disables syntax checking in ACE. If we want to use syntax
checking in the future, we can set "useWorker: true" and import the
"ace-builds/webpack-resolver" file.