[app] Fix Permission Handling for Applications (#413)
This commit fixes the permission handling for applications, when
authentication is enabled, but the user is not part of any teams. In
this case a user was still able to see all applications. We changed the
implementation to handle such cases like the user wants to see all
applications so that the access check is executed.