[jira] Add Jira Plugin
The Jira plugin can be used to access issues from a Jira instance in
kobs. The user must provide an email address and API token to use the
plugin. The plugin can be used in panels and in the notification, e.g.
to show the users active sprint or the issues he is assigned to.