[jira] Convert description to markdown (#415)
The description of an issue is now converted to markdown before we
render the description. This is necessary because Jira returns the
description in a wiki markup format, which doesn't look good when it is
directly passed to the ReactMarkdown component for rendering.

We also changed the logic to display the details of an issue, so that
only the fields which are relly available are displayed in the details.
This means if for example an issue doesn't have subtasks we do not
display the subtasks section anymore.

We also changed the color of the status label, so that the label matches
the color in Jira. The previously used color field seems to be
deprecated and Jira only allows 3 status categories. For these
categories we set the color: "Done" -> "green", "In Progress" -> "blue"
and "Open" -> "grey".