[app] Fix error handling in resources table (#416)
This commit fixes the error handling in the resources table. When an
error was returned and a valid list of resources, we did not show the
list, but instead we only show the errors.

Now we are checking if there is a valid list of resources. If this is
the case we render the table and ignore the errors.

We also fixed the handling of links in notification. Until now we always
displayed an external link icon which was used to redirect a user to the
details. Now the user can directly click on the notification to get
redirected to the details for a notification.