[jira] Adjust notification style (#417)
We are now using different styles for the issues displayed in the
notification drawer. The style is based on the status of an issue. When
an issue is in the status "Open" the notification item is yellow, for
"In Progress" the notification item is blue and for "Done" is is green.
For issues which are "Done" the notification is also marked as read.