[signalsciences] Add Signal Sciences Plugin
This commit adds a new plugin for Signal Sciences. The plugin allows
users to view the statistics of a site, the agents for each site and
users can search for requests.

Note: Currently we are alsways retrieve 1000 requests from the Signal
Sciences API and handling the pagination on the client. When
https://github.com/signalsciences/go-sigsci/pull/28 gets merged we can
change this to handle the pagination within the API requests to reduce
the number of requests which are returned and to to allow users to view
more then 1000 requests for their queries.