[app] Use generics for jwt package
We are now using generics for the jwt package, so that we can reuse the
logic for cookies which are generated in plugins.

We use the jwt package now in the GitHub and Jira plugin. This has the
advantage that the cookies are more secure now. Before this change we
only used a base64 encoded representation of the auth credentials. Now
we generating a signed JWT token to store the auth info.