[app] Rework "useDimensions" hook (#422)
We changed the "useDimensions" hook, so that the hook now also takes
into account when the sidebar is toggled. Before the rework the returned
height and width of the "useDimensions" hook was not changed when a user
opens/closes the sidebar, so that some components where not correctly
adjusted. This is now fixed by using the "getResizeObserver" from the
"@patternfly/react-charts" package within the hook.