Add Release Drafter
We are now using the Release Drafter to generate the changelog for new
releases. For that reason we are also removing the CHANGELOG.md file, so
that users do not have to add manuall entries to the changelog.

The changelog is generated based on the following labels: "changelog:
added", "changelog: fixed" and "changelog: changed". One of these labels
must be added to a PR.