[sql] Fix query encoding in panel actions
The query parameter was not encoded in the panel actions, so that the
a users query broke, when he selected a query in the action and was
redirected to the raw view of his data.

This is now fixed by using "encodeURIComponent" for the query parameter
in the actions of an sql panel.