[jaeger] Add Service Performance Monitoring (#428)
It is now possible to view the Jaeger Metrics for Service Performance
Monitoring within kobs.

Service Performance Monitoring allows users to view the latency, error
rate and request rate for all operation of a service. It is also
possible to switch from the monitor view directly to the corresponding
traces view of a selected operation.