[jaeger] Add Panels for Service Performance Monitoring
It is now possible to use the new Service Performance Monitoring feature
of the Jaeger plugin within dashboards. It is possible to add panels for
the service latency, error rate and request rate and a panel for the
metrics of the operations of a service.

It is now also possible to select the span kinds which should be
included in the different metrics. By default we always include all span
kinds.