[mongodb] Fix null error and limit parameter
When a field in MongoDB is "null" the plugin throws an error. This
should be fixed now, by only calling the "toString()" method, when the
field is not "null".

This commit also fixes an error with the limit parameter, which was not
used correctly, because it was not added in the "changeOptions"
function. We also improved the form in the query toolbar to prevent the
default submit action.