[klogs] Use maps instead of columns for custom fields
This is a follow up for https://github.com/kobsio/klogs/pull/34 were we
changed the database schema for klogs to use maps instead of arrays for
custom fields.

To be able to query the data ingested by the klogs plugin we had to
adjust the queries to also use maps instead of columns for these fields.