[mongodb] Fix render performance (#439)
The MongoDB plugin had some performance issues when it rendered a large
list of collection or large documents. The performance issues for the
collections lists are fixed by adding pagination to the collections
panel. The performance issues for large documents are fixed by limiting
the number of lines for the editor view to 50 and by only rendering the
document parts which are expanded.