[mongodb] Fix pagination for collections (#440)
There was a small bug in the pagination logic for the collections
overview. Instead of the count of the filtered collections we always
used the complete count of the collections in the pagination logic. This
is now fixed.

We also changed the logic to reset the users selected page, when he
changed the filter.

Last but not least we now show the collection name on the queries and
document page besides the plugin name and satellite name.