[app] Extend Placeholders for Dashboards
It is now possible to set a default value for a placeholder in a
dashboard and it is possible to set a type for the placeholder.

Until now, it was only possible to use string values for placeholders.
This is now changed so that also "numbers" and "objects" can be used.