[app] Fix User Authentication
For users with a lot of teams and permissions the authentication process
was broken, because we saved all the users teams and permissions within
the cookie, so that the cookie could exceed the maximum size of 4096
characters.

This is now fixed by just adding the users email address and team to the
cookie and getting the permissions within each API call via the auth
middleware.