[core] Adjust launch.json for VS Code
This commit removes the ".env" configuration in the VS Code launch.json
for the Go commands, so that we can run the command only with a
"config.yaml" file in the root folder.

This also adjusts the "app" command so that we can debug the React app
via VS Code and the Chrome Debugger.