[core] Adjust Dashboards
Adjust dashboards layout to use MUI instead of Patternfly and to use
react-grid-layout, so that panels in a dashboard can be dragged and
resized via the app in the future.

The dashboards page was adjusted too, so that wen can now display
multiple dashboards within the dashboards page. A dashboards page can
now also be linked from the default / user provided navigation.