[core] Adjust Applications and Teams Page (#469)
Adjust the applications and teams page for the new design. This commit
adjusts the pages to show a list of applications, a single application,
a list of teams, a single team and a topology graph to use the new
design defined by the rework done in #454.

This commit also adds some utility components which also should be used
by other plugins:

- DetailsDrawer: To display some details of a resource within a drawer
- Page: Defines the general layout of a React Router route
- Pagination: Paginate through query results, where the number of items
  per page can be selected
- Toolbar: A toolbar layout which can be displayed within a Page
  component
- UseQueryWrapper: Handle all possible states returned by useQuery

Next to the components this commit also improves the existing hooks and
adds some new hooks:

- useDebounce: Debounce changes to a value
- useDimensions: Get the width and height of an element
- useIsWidthUp: Check if the width of the window is larger then a
  specific breakpoint
- useQueryState: Manage the state of the app via url query parameters

Note: Some features are not implemented yet. This will be done in a
follow up PR, these features are:

- Show dashboards on the application and team page
- Show insight charts of an application
- Add / edit applications and teams