[core] Add Panels for Applications, Teams and Dashboards
Add the "core" panels for applications, teams and dashboards. In detail
these are the following panels:

- "applications": Display a list of applications for a provided team or
  for the currently authenticated user.
- "applicationgroups": Group the applications for a provided team or for
  the currently authenticated user.
- "topology": Display the topology graph for a single application.
- "teams": Display a list of teams for the currently authenticated user.
- "dashboards": Display a list of links to dashboard pages. Each page
  can contain multiple dashboards. This is similar to how we handle
  links to dashboards in the navigation.

To display these panels within a dashboard the new "PluginPanel"
component is used. This component should also be used by other plugins
to ensure a unique styling across the app.

The second "PluginPanl" component (the one in the "plugins" folder), is
used to load the correct plugin based on the user provided options for a
panel. If the type of the plugin is "core" we are loading the core
panels mentioned above.

The rows in a dashboard are now containing a new "autoHeight" property,
which can be used to adjust the height of each panel in a row based on
the content. This will not work with all plugins and should mainly be
used for rows which are just containing one panel.