[core] Fix Applications and Teams (#481)
The application and team page were still showing the "TODO: Show
Dashboards" message, which was already implemented, but we forgot to
remove this placeholder.

The applications panel used the used the useQueryState hook for the page
and per page state, wiche caused that all panels were rerendered when a
user select a new page in the panel. This is now fixed by using the
useState hook to manage the state.