[core] Improve Editor Component
This commit adds a "lineNumbers" and "minimap" property to the Editor
component, which allows us to disable the line numbers and minimap in
the editor. This is currently not required, but necessary for later when
we want to use the Editor as "normal" text field.

This commit also improves the setup and build of monaco, which is used
as the underlying editor for our Editor component, so that we are now
able to register custom plugins (like the one for yaml) and do not have
to rely on a CDN to load the editor.

The currently supported Nord theme (https://www.nordtheme.com, you
should use it, it is awesome^^) was also improved, by converting the VS
Code theme to a monaco theme and by adding some custom tokens for yaml.

We also added the monaco-yaml plugin, to validate the syntax of yaml
files. This also allows us to add some schemas for the Kubernetes
resources later, which can be used for additional validation.

Finally the build setup of Vite was slightly adjusted so that our css
files are now included in the production build (before these changes
they were missing which resulted in some strange layouts for the
terminal and the grid in dashboards). These adjustments were also
required to add the workers from monaco to the production build of the
app package.