[core] Fix Positioning of Options Component in Toolbar
Instead of using the ToolbarItem directly in the Options component, the
ToolbarItem wrapper should be set in the component which uses the
Options, so that we can apply the correct ToolbarItem props.

This is required, because it was not possible to position the Option
component in the toolbar next to a ToolbarItem, which already used the
"grow" property.