[core] Fix Font Family for Monaco (#485)
Monaco should use the same font family as we use in other places in the
app for monospaced text. MUI and xterm are using "monospace" as font
family, so monaco should use the same.

Note: If we find a better font family, we can also change this in the
future.