[helm] Adjust Helm Plugin
This commit adjusts the Helm plugin to use the new design and plugin
architecture. This means that the Helm plugin must only be registered at
the "hub" and not on each cluster anymore. The permissions must also not
specified anymore, instead the resources permissions will be used with
the "secrets" resource.