[helm] Add Panel Actions
This commit adds an "Explore" action to the Helm panels, which allows a
user to go to the Helm plugin page, from a dashboard panel. The provided
clusters and namespaces are added as query parameters.

We also replaced the Paper component with a Card component to remove the
bottom border which looks better when the plugin is used in a dashboard.