[core] Fix DetailsDrawer and Replace Paper Component (#501)
This commit fixes the actions shown in the DetailsDrawer which were not
aligned to the right side of the drawer. This is now fixed by adding the
"flexGrow" property to the correct component.

This commit also replaces the Paper component on the sign in page and
the resources panel with a Card component, so that we do not have a
bottom boarder, which looks a bit nicer.