[prometheus] Add Query History
All queries which are executed by a user are now added to the state
history, so that a user can always access the latest queries he
executed. The history is saved in the localStorage of the users browser.
To reuse the logic some new helpers where added, these are:

- "getStateHistory": to return the history for a provided key from the
  localStorage.
- "addStateHistoryItem": to add a single item to the history
- "addStateHistoryItems": to add a list of items to the history, before
  the items are added, empty items are filtered and the items are
  deduplicated, then we store the 10 latest items in the localStorage

This commit also fixed a bug in the "Editor" and "MUIEditor" component,
where we used the "defaultValue" and "defaultLanguage" properties, so
that the value was not updated in the monaco editor when it was changed
in another component (e.g. value from the history). Now we are using the
"language" and "value" properties so that the value is always updated.