[core] Add "Clear" Action to Search Fields (#505)
This commit adds a clear action to all search fields, which should make
it easier for users to delete a provided search term. The clear button
is displayed within the TextField component via the input adornment API.

We also adjusted the query fields in the Prometheus plugin, so that the
actions are also displayed within the TextField and not next to it, so
that we have a similar style across the app.