[prometheus] Reduce "getStateHistory" Calls
The number of calls of the "getStateHistory" function is now reduced by
wrapping the call in the "useMemo" hook. This also reduces the calls to
the browsers localStorage. The "options.queries" are added as dependency
to the useMemo hook, so that the history is refreshed each time a user
executes a query / a new item is added to the history.