[core] Persist User Selected Options (#502)
The options which are selected by a user in a toolbar are now persisted,
so that the user will can reuse the options when he navigates a second
time to the same page. To achieve this we had to adjust the
"useQueryState" hook, so that we can overwrite the default value of an
array with an empty array. We also introduced a new hook
"useLocalStorageState" which is used to persist the option in the
browsers localStorage.