[sonarqube] Change API Endpoint to Get Projects
Until now we used the "api/projects/search" API endpoint of SonarQube to
get a list of projects. This is now changed to "api/components/search"
because this endpoint requires less permissions for the provided token
and also works with a normal token.

We also added "unmountOnExit" property to the accordion in the projects
list, so that the measures for a project are only loaded when the
accordion is opened.