[core] Allow Plugin Mounts to Return Nil
The "MountCluster" and "MountHub" function can now return "nil" for the
router. This way we do not have to specify a not found route when the
plugin can not be used within a "cluster" or "hub".

The following plugin were also adjusted, so that they can only be used
within the "hub":

- Harbor
- Opsgenie
- RSS
- SignalSciences
- SonarQube