[core] Add Plugin Permission Check
This commit re-adds the permission check for plugins. We already checked
if a user is allowed to access a plugin in former versions of kobs, but
the check was migrated to the new plugin architecture yet. With this
change the permission check is now also present in the new version and
only allows users with the correct rights to access a plugin.

This commit also adjusts the Flux and Helm plugin to correctly use the
permissions set for a user. Some parts of the Flux plugin can also be
used without access to the plugin, because the "/api/resources" endpoint
is used, for these requests the user must have the permissions to view
resources on the mentioned "/api/resources" endpoint.

Last but not least the plugins returned by our API are now filtered
before returned, so that a user only sees the plugins he has access to
in the frontend.