[core] Fix Time Handling in "useQueryState" Hook (#514)
This commit fixes the handling of the "time", "timeEnd" and "timeStart"
query parameters in the "useQueryState" hook. Instead of using the times
directly we have to check if the "time" parameter contains a valid time
option and have to set the corresponding times if this is the case.

This way the time parameters are now handled as in former version via
the "getTimeParams" function, which was is also removed now, since it
shouldn't be used anymore (it also wasn't used yet).