[core] Improve Padding for Plugin Panels
The padding applied to the content of a plugin panel was missing when
the provided content was scrollable, this is now fixed, so that there is
always a bottom padding for the panels, so that the look better within a
dashboard.

Because of the changes in the PluginPanel component, we also had to
adjust the Prometheus Plugin so that there is no scrolling in scrolling
applied to the charts.

This commit also fixes the applications group panel, where the divider
was also rendered after the last element in the list, which was not
intended, so that the last divider is now removed.