[prometheus] Add Sorting for Tables
The data shown in a Prometheus table can now be sorted by each column.
For that we convert the returned records from the API to a list of rows,
based on the user provided columns. The returned rows can then be sorted
in a ascending descending order and by the values of each column.