Prometheus use ChartTooltip Component (#517)
Use the ChartTooltip component which was added within the Jaeger plugin
(#516, ok actually it was added in a slightly different form in #472) in
the Prometheus charts. This allows us to remove the custom ChartTooltip
component.

The ChartTooltip component is to prefer, because it handles the tooltip
alignment in a better way and we can ensure that all charts have a
similar tooltip.

This commit also renames the "label" property to "name", because we can
not use "label" as property in a data point. If we use "label" as
property in a datapoint this will automatically render a label for each
datapoint in the chart, which is not wanted.