[kiali] Adjust Kiali Plugin (#519)
Adjust the Kiali plugin to use new plugin architecture and design.

The Kiali plugin can now also be used to render the topology graph for a
single application, when it is used within a dashboard. To do this one
namespace must be provided in the "namespaces" options and the name of
the application must be provided via the new "application" option.