[jaeger] Add Flamegraph
Add flamegraph visualization for graphs, this means a user can now
switch between the timeline and flamegraph visualization via the trace
actions. Similar to the official Jaeger UI we are using the
"@pyroscope/flamegraph" package to render the flamegraph.

This commit also fixes a small bug in the traces view, when a trace
contains a lot of services, which are now displayed on multiple lines to
avoid the horizontal scrolling.