[core] Replace Monaco with CodeMirror
This commit replaces Monaco with CodeMirror in our Editor component,
because we come across several problems with Monaco. For example it was
not possible to render multiple instances with different themes, because
of the usage of globals in Monaco.

CodeMirror feels a bit more modern and doesn't have these problems, so
that we want to use it instead of Monaco.

Besides the rewrite of the Editor component the following changes were
done:

- Remove all setup related things, which were only added for Monaco
- Do not limit the height of the rendered Editor component, since the
  new Editor automatically adjusts it's height
- Rewrite the custom autocompletions / languages for CodeMirror
- Add a new "/proxy" API route for Prometheus, so that we can use the
  "@prometheus-io/codemirror-promql" package in the Prometheus plugin