[core] Exclude "index.ts" File from Coverage Report (#529)
The "index.ts" file shouldn't be included in the test coverage resport,
so it is added as exclude. Since this seems to add some other files to
the report we now define that only files in the "src" folder are
included in the coverage report (and exlude "index.ts" and
"vite-env.d.ts").