[core] Add Application Insights
As known from former versions of kobs this commit re-adds the
application insights feature. Insights can be used to render sparkline
charts for the most important metrics within an application and normally
displayed via the details drawer of an application.

The insights can also be used as panel on a dashboard via the
"applicationinsights" plugin ("type=core"), to render the insights of an
application in a dashboard panel. For this a new option "direction" was
added, which can be used to show the insights in a "row" or "column".