[mongodb] Fix Overflow in Documents View
It could happen that the table layout breaks in some case causing a
vertical scroll in the documents view. This is now fixed by breaking all
words.

The table cells are now also not vertically aligned at the top, because
this doesn't look that good, so that they are now vertically aligned in
the middle.