[techdocs] Fix Vertical Scrolling in Service Pages (#535)
When a code block contained very long lines it could happen that a user
must scroll vertically on a service page. This is now fixed by wrapping
long lines in a code block.

Because of this new setting copying content from a code block adds some
useless newlines, so that we added a copy button to each code block to
make it easier for users to copy the content.