[klogs] Fix Add Filter for Number Fields
When a user used the add filter buttons from the table view of an
document to search for a value or to exclude a value this was only
working for fields of type string, because we didn't know the type of a
field in the React app.

This is now fixed, by returning an array of objects from our API with
each field as "name" property and the field type as "type" property.
This information is then used in the React app to adjust the filter
which should be added based on the type of the field.