[core] Add / Edit CRs via the Frontend (#543)
It is now possible to add / edit new and existing CRs for Applications,
Users and Teams via the frontend. Therefor the Add / Edit button on the
corresponding resource page can be used.

This feature is disabled by default and must be enabled via the
configuration file ("hub.app.settings.save.enabled"). Changes made by a
user for a CR which is managed via the watcher are overwritten on each
reconciliation.