[jira] [github] Improve Authentication Handling
This commit improves the authentication handling for the Jira and GitHub
plugin, so that a user do not have to re-authenticate that often.
Therefor we increased the default time for which the authentication
cookie is valid to one month and re-create the cookie with every call to
the auth endpoint. To ensure that the endpoint is called on a regular
basis the stale time for these queries is set to 24 hours.

The cookies for the plugins are now also deleted when the user clicks on
the signout button.