:warning: [core] Adjust Helm Chart
Instead of providing a seperate Helm chart for the "hub" and "satellite"
component, we decided to create a single Helm chart named "kobs" which
can be used to deploy all components. The components can be enabled /
disabled sperately in the Helm chart, so that it is still possible to
just deploy one component using the Helm chart.

We also adjusted the configuration for all components to make it easier
to configure them via the Helm chart. This means that there are no
"gloabl" settings anymore an everything must be configured on a per
component basis (e.g. instead of "log.level" the log level can now be
configured via the "<component>.log.level" option).