[core] Adjust Kustomize Files (#553)
Adjust the Kustomize files, so that the kobs components: hub, cluster
and watcher can be deployed via Kustomize. The Kustomize files are the
rendered Helm chart with the default settings.

Fix the naming of the ClusterRole, ClusterRoleBinding and ServiceAccount
for the cluster component in the Helm chart (remove the duplicated dash
in the name).

Adjust Pull Request Template to point to the correct values.yaml file
for our Helm chart.