[core] Fix TeamPage
The useQuery parameter in the TeamPage component was wrong, instead of
the "id" we were using the "cluster", "namespace" and "name" parameters.
This is now fixed.

Also the divider shown in the teams list was fixed, so that the divider
is now displayed in the same cases as in the applications page.