[core] Adjust DetailsDrawer (#566)
In some places we didn't had a padding on the Cards displayed within the
DetailsDrawer component. This is now fixed, so that all Cards should
have a padding of the same size.

The DetailsDrawer was always mounted in some components, instead of just
mounted when it is open. This was also fixed, so that the DetailsDrawer
is only mounted when displayed.