[core] Fix Save Function when Save is Disabled (#569)
When the "save" options is disabled a user should still be able to edit
the yaml file of an application, team or user directly via the frontend.
Until now this wasn't working because the changes made to a resource in
the save preview field only changed the resource when the save option
was enabled. This is now changed so that the yaml file can also be
adjusted when the save option is disabled.