[core] Fix Editor (#573)
The fix for CodeMirror #568 only worked partially, it fixed the
production build but broke the autocompletion and syntax highlighting
for some languages. To fix it we decided to move the complete language
logic to the [core] plugin. This means in the plugins we only specify
the "language" (e.g. "mongodb", "sql", "klogs", ...) and some
"languageOptions" which are specific for each language. The
configuration of this language for the editor then happens in the core
package.