[core] Fix Dashboard Updates for Back Button
When a user jumps from one application to another one using the topology
tags on the application page and then uses the back button to go back to
the original application the dashboards view wasn't updated.

This is now fixed by updating the variables from the dashboard property
in the "Dashboard" component using "useEffect".