[runbooks] Add Runbooks Plugin
The new runbooks plugin can be used to manage runbooks for alerts via
kobs and Custom Resource Definitions which are compatible with the
PrometheusRule CRD (e.g. VMRules).

The runbooks for each alert can be provided via a "runbook" annotation
within these CRs. Besides the "runbook" annotation the plugin also uses
the "message" annotation and the "severity" label from these CRs.

The CRD which should be used in the plugin can be configured via the
"path" and "resource" field in the plugin options.

Besides the added runbook plugin, this commit also fixes the "Editor"
component, where the "readOnly" mode wasn't working, so that we had an
"readOnly" propertie to explicit set the editor in a read only mode.