[core] Add Debug Endpoints
This commit adds some debug endpoints which are provided via an
additional debug server, which listens on port 15225 by default and can
be enabled in the configuration file via the "debug.enabled" property.

The debug endpoints containing the pprof endpoints and two custom
endpoints for testing request timesouts and dumping a whole request.